Get the tests working neatly
