+++ python_client/.gitignore
@@ -1,0 +1,1 @@
+oauth-credentials.json
