--- python_client/tests/cassettes/repo.json
@@ -1,97 +1,0 @@
-{
-  "http_interactions": [
-    {
-      "response": {
-        "url": "https://api.wellcomecollection.org/storage/v1/ingests/017e60b8-9779-463f-821e-321a20f7edc4",
-        "headers": {
-          "Content-Type": [
-            "application/json"
-          ],
-          "Access-Control-Allow-Credentials": [
-            "true"
-          ],
-          "x-amzn-Remapped-Content-Length": [
-            "1056"
-          ],
-          "Via": [
-            "1.1 089ee28df029aa2f4878a528c124aaf9.cloudfront.net (CloudFront)"
-          ],
-          "Access-Control-Allow-Methods": [
-            "GET, POST, OPTIONS"
-          ],
-          "Access-Control-Allow-Origin": [
-            "*"
-          ],
-          "x-amz-apigw-id": [
-            "W8s9mHmijoEFTDg="
-          ],
-          "x-amzn-Remapped-Connection": [
-            "keep-alive"
-          ],
-          "Date": [
-            "Fri, 22 Mar 2019 14:45:56 GMT"
-          ],
-          "X-Amz-Cf-Id": [
-            "u6BQWsYwgcDjI5iMn90u7saQ3llwD4zLbW_2PqL99H20VKDK7lgS2A=="
-          ],
-          "Content-Length": [
-            "1056"
-          ],
-          "x-amzn-Remapped-Date": [
-            "Fri, 22 Mar 2019 14:45:56 GMT"
-          ],
-          "Access-Control-Allow-Headers": [
-            "DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type"
-          ],
-          "x-amzn-Remapped-Server": [
-            "nginx"
-          ],
-          "X-Cache": [
-            "Miss from cloudfront"
-          ],
-          "x-amzn-RequestId": [
-            "32ee9726-4cb1-11e9-82a1-1df7e6ea9574"
-          ],
-          "Connection": [
-            "keep-alive"
-          ]
-        },
-        "body": {
-          "string": "{\"@context\":\"https://api.wellcomecollection.org/context.json\",\"id\":\"017e60b8-9779-463f-821e-321a20f7edc4\",\"sourceLocation\":{\"provider\":{\"id\":\"aws-s3-standard\",\"type\":\"Provider\"},\"bucket\":\"temp-bagit-drop-test\",\"path\":\"b16784261.zip\",\"type\":\"Location\"},\"ingestType\":{\"id\":\"create\",\"type\":\"IngestType\"},\"space\":{\"id\":\"digitised\",\"type\":\"Space\"},\"status\":{\"id\":\"succeeded\",\"type\":\"Status\"},\"bag\":{\"id\":\"digitised/b16784261\",\"type\":\"Bag\"},\"events\":[{\"description\":\"Started work on ingest: 017e60b8-9779-463f-821e-321a20f7edc4\",\"createdDate\":\"2019-01-16T01:27:22.324180Z\",\"type\":\"IngestEvent\"},{\"description\":\"Ingest bag file downloaded successfully.\",\"createdDate\":\"2019-01-16T01:27:23.023765Z\",\"type\":\"IngestEvent\"},{\"description\":\"Bag uploaded and verified successfully\",\"createdDate\":\"2019-01-16T01:27:27.642623Z\",\"type\":\"IngestEvent\"},{\"description\":\"Bag registered successfully\",\"createdDate\":\"2019-01-16T01:27:27.983454Z\",\"type\":\"IngestEvent\"}],\"createdDate\":\"2019-01-15T13:26:46.080864Z\",\"lastModifiedDate\":\"2019-01-15T13:26:46.080864Z\",\"type\":\"Ingest\"}",
-          "encoding": null
-        },
-        "status": {
-          "code": 200,
-          "message": "OK"
-        }
-      },
-      "request": {
-        "method": "GET",
-        "uri": "https://api.wellcomecollection.org/storage/v1/ingests/017e60b8-9779-463f-821e-321a20f7edc4",
-        "headers": {
-          "Accept-Encoding": [
-            "gzip, deflate"
-          ],
-          "User-Agent": [
-            "python-requests/2.21.0"
-          ],
-          "Accept": [
-            "*/*"
-          ],
-          "Authorization": [
-            "<AUTH_TOKEN>"
-          ],
-          "Connection": [
-            "keep-alive"
-          ]
-        },
-        "body": {
-          "string": "",
-          "encoding": "utf-8"
-        }
-      },
-      "recorded_at": "2019-03-22T14:45:56"
-    }
-  ],
-  "recorded_with": "betamax/0.8.1"
-}
