+++ python_client/tox.ini
@@ -14,3 +14,8 @@ basepython = python2.7
 deps = pip-tools
 commands =
     pip-compile test_requirements.in
+
+[testenv:lint]
+basepython = python2.7
+deps = flake8
+commands = flake8 --max-complexity 10 src tests
